(#1016) move nonNuls from input,output,bytes to own packages
